CASMTRIAGE-6065 Use safer script
The `create-kis-artifacts.sh` script is not safe for fully deployed
systems.

Use `create-ims-initrd.sh`, this will not disrupt the running system.

We do need to comment out the lines in the `create-ims-initrd.sh` script
that remove the kdump initrd, since we may end up needing that if
catastrophy strikes and the system panics while running the hotfix.
After the script is called, this restores the original copy of the
script.

Also adds a lot more logging, as well as non-colliding backups of BSS.