Fix order of operations
The code for obtaining and using `CSM_CONFIG_COMMIT` is running before `setup-nexus.sh` has loaded the new csm-config. This code fails as-is because the commit can not be resolved until csm-config is loaded in by `setup-nexus.sh`. This was missed in testing due to constant re-runs, where `CSM_CONFIG_COMMIT` had been set in the developer's environment.

Likewise, loftsmand can not deploy the chart until `setup-nexus.sh` has loaded the chart into Nexus.