CASMTRIAGE-7232 CSM 1.5 kdump hotfix
This adds a hotfix for kdump in 1.5.

In some version of 1.5, namely 1.5.2, kdump will emit the following
error during system startup:
```
Cannot reserve 256MB crashkernel low memory, please try smaller size.
```
The result is no core dump is taken during a system crash.

The error is new as of a minor version bump of kdump, but the changelogs
were inconclusive.

The fix was to simply set a `low` memory bound for kdump. After setting
this on the commandline the system is able to take core dumps again.